phpunitでテストコードを作成する

+++ src/app/Http/Controllers/ApiCategoryController.php
@@ -4,15 +4,17 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Category;
+use App\Http\Requests\CategoryRequest;
 use Illuminate\Support\Facades\Auth;
 
 class ApiCategoryController extends Controller
 {
   public function getAll()
   {
-      $categories = Category::all();
+      $userId = Auth::id();
+      $categories = Category::where('user_id', $userId)->get();
 
-      if (!$categories) {
+      if ($categories->isEmpty()) {
         return response()->json([
             'message' => 'categoryが見つかりません'
         ], 404);
@@ -23,19 +25,21 @@ class ApiCategoryController extends Controller
       ], 200);
   }
 
-  public function create(Request $request)
+  public function create(CategoryRequest $request)
   {
       $categoryInput = $request->all();
       $name = $categoryInput['name'];
+      $userId = $categoryInput['user_id'];
 
       Category::create([
         'name' => $name,
+        'user_id' => $userId,
       ]);
 
       return response()->json($categoryInput);
   }
 
-  public function updateById(Request $request, $id)
+  public function updateById(CategoryRequest $request, $id)
   {
       $category = Category::find($id);
 
@@ -57,16 +61,34 @@ class ApiCategoryController extends Controller
             'message' => 'categoryが見つかりません'
         ], 404);
       }
+
+      $authUserId = Auth::id();
+      $userId = $category['user_id'];
+      if ($authUserId !== $userId) {
+        return response()->json([
+            'message' => '認証に失敗しました'
+        ], 401);
+      }
+
       $category->delete();
       return response()->json($category);
   }
-  public function getById(Request $request, $id)
+
+  public function getById($id)
   {
       $category = Category::find($id);
       if (!$category) {
         return response()->json([
             'message' => 'categoryが見つかりません'
         ], 404);
+      }
+
+      $authUserId = Auth::id();
+      $userId = $category['user_id'];
+      if ($authUserId !== $userId) {
+        return response()->json([
+            'message' => '認証に失敗しました'
+        ], 401);
       }
 
       return response()->json([
